refactor: move implement block
