+++ src-tauri/src/api/auto_partition.rs
@@ -5,40 +5,6 @@ use std::vec::Vec;
 use std::{clone, str::FromStr};
 use tea_partition_generator::single_boot_blockdev::{Blkstuff, SingleBootBlockdevice};
 
-impl From<tea_partition_generator::blueprint::Partition> for crate::installer::Partition {
-    fn from(data: tea_partition_generator::blueprint::Partition) -> Self {
-        crate::installer::Partition {
-            number: data.number,
-            disk_path: data.disk_path,
-            path: data.path,
-            mountpoint: data.mountpoint,
-            filesystem: data.filesystem,
-            format: data.format,
-            start: data.start,
-            end: data.end,
-            size: data.size,
-        }
-    }
-}
-
-fn convert_partition(
-    opt: Option<Vec<tea_partition_generator::blueprint::Partition>>,
-) -> Option<Vec<crate::installer::Partition>> {
-    opt.map(|vec| vec.into_iter().map(Into::into).collect())
-}
-
-impl From<tea_partition_generator::blueprint::Storage> for crate::installer::Storage {
-    fn from(data: tea_partition_generator::blueprint::Storage) -> Self {
-        crate::installer::Storage {
-            disk_path: data.disk_path,
-            partition_table: data.partition_table,
-            new_partition_table: data.new_partition_table,
-            layout_changed: data.layout_changed,
-            partitions: convert_partition(data.partitions), 
-        }
-    }
-}
-
 #[tauri::command]
 pub async fn autogen_partition_select_disk(
     blkname: String,
@@ -48,11 +14,11 @@ pub async fn autogen_partition_select_disk(
 ) -> () {
     // Handle singleboot
     if mode == "singleboot" {
-        let ctx: Blkstuff = SingleBootBlockdevice::blockdevice(
-            "/dev/sdb".to_string(),
-            "btrfs".to_string(),
-            "gpt".to_string(),
+        println!(
+            "generating config {} for {} with fs {} -> {}",
+            mode, blkname, fs, partition_table
         );
+        let ctx: Blkstuff = SingleBootBlockdevice::blockdevice(blkname, fs, partition_table);
 
         let ret = ctx.getresult();
 
@@ -60,7 +26,7 @@ pub async fn autogen_partition_select_disk(
             // Start werite json
             let mut blueprint = super::get_blueprint().unwrap();
 
-            blueprint.storage = Some(ret_val.into()); 
+            blueprint.storage = Some(ret_val.into());
 
             super::write_blueprint(blueprint).unwrap();
         }
