+++ src-tauri/src/utils/mod.rs
@@ -1,1 +1,2 @@
 pub mod command;
+pub mod tea_partition_generator_convert;
